Add files via upload
Volunteers participating in the experiment were taken to the corresponding experimental sites without knowing the content of the experiment. Then, the volunteers were led from the designed route through the corresponding experimental mall, and after reaching the endpoint, the volunteers were asked to write down the memories of the name of the POI stores in the mall. The process of volunteers recalling POI stores in experimental sites is considered to be a process of users' cognitive space.